Rename sp_vm_function to scripted-invoker.
