+++ sourcepawn/jit/scripted-invoker.h
