+++ sourcepawn/jit/sp_vm_engine.h
@@ -14,7 +14,7 @@
 #define _INCLUDE_SOURCEPAWN_VM_ENGINE_H_
 
 #include "sp_vm_api.h"
-#include "sp_vm_function.h"
+#include "scripted-invoker.h"
 
 class BaseContext;
